+++ jsonschema2pojo-gradle-plugin/pom.xml
@@ -18,7 +18,7 @@
     <dependencies>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>jsonschema2pojo-cli</artifactId>
+            <artifactId>jsonschema2pojo-core</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
